requirements.txt
fastapi~=0.90.0
uvicorn
sqlalchemy~=2.0.3
passlib~=1.7.4
bcrypt
pydantic~=1.10.4
python-multipart # to process Form data from the frontend 
python-jose #for jwt encoding